feat(2FA): [PM-17129] Login with 2FA Recovery Code - Feature flagged implementation.

+++ src/Identity/IdentityServer/RequestValidators/BaseRequestValidator.cs
@@ -181,10 +181,15 @@ public abstract class BaseRequestValidator<T> where T : class
 
             // 3d. If the user is logging in using a RecoveryCode type to login
             // as their second factor we know we are in the flow where we need
-            // to disable their other 2FAs.
-            if (twoFactorProviderType == TwoFactorProviderType.RecoveryCode)
+            // to disable their other 2FAs just like as if they were using
+            // their recovery code because they no longer have access to their
+            // 2FA device.
+            if (FeatureService.IsEnabled(FeatureFlagKeys.RecoveryCodeLogin))
             {
-                await _userService.RemoveTwoFactorProviderAsync(user);
+                if (twoFactorProviderType == TwoFactorProviderType.RecoveryCode)
+                {
+                    await _userService.RemoveTwoFactorProviderAsync(user);
+                }
             }
         }
